% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitVg.R
\name{fit.vG}
\alias{fit.vG}
\title{Fits van Genuchten Model}
\usage{
fit.vG(theta, P, theta.s = -999, theta.r = -999, alpha = -999, n = -999)
}
\arguments{
\item{theta}{moisture content cm_3}

\item{theta.s}{saturated moisture content cm^3^ cm^{-3}^}
}
\description{
Fits van Genuchten Model
}
\examples{

}
\keyword{conductivity}
\keyword{hydraulic}
\keyword{retention}
\keyword{saturated}
\keyword{water}
