% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mualem.R
\name{Mualem}
\alias{Mualem}
\title{A Function}
\usage{
Mualem(P, alpha, n, Ko, L)
}
\arguments{
\item{parameters}{here.}
}
\description{
A Function
}
\examples{

}
\keyword{conductivity}
\keyword{hydraulic}
\keyword{retention}
\keyword{saturated}
\keyword{water}
