% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoreDF.R
\name{PoreDF}
\alias{PoreDF}
\title{Pore DF function}
\usage{
PoreDF(r, alpha, n)
}
\arguments{
\item{r}{parameter}

\item{alpha}{parameter}

\item{n}{parameter}
}
\description{
Pore DF function
}
\examples{

}
\keyword{keywords}
\keyword{some}
