% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.vG.R
\name{predict.vG}
\alias{predict.vG}
\title{Calculates an Genuchten Model based on parameters requires Van Genucten equation (vG) as well}
\usage{
\method{predict}{vG}(P, fit)
}
\arguments{
\item{P}{- is pressure head}

\item{fit}{- data frame with theta.s theta.r alpha and n}
}
\description{
Calculates an Genuchten Model based on parameters requires Van Genucten equation (vG) as well
}
\examples{

}
\keyword{conductivity}
\keyword{hydraulic}
\keyword{retention}
\keyword{saturated}
\keyword{water}
