% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectiveSaturation.R
\name{SE}
\alias{SE}
\title{A Function to calculate effective saturation}
\usage{
SE(theta, theta.s, theta.r)
}
\arguments{
\item{theta}{-- volumetric water content (cm^3 cm^(-3))}

\item{theta.s}{-- saturated volumetric water content (cm^3 cm^{-3})}

\item{theta.r}{-- residual water content (cm^3 cm^(-3))}

\item{SE}{-- effective saturation (unitless) plugs into water retention functions}
}
\description{
A Function to calculate effective saturation
}
\examples{


}
\keyword{conductivity}
\keyword{hydraulic}
\keyword{retention,}
\keyword{saturated}
\keyword{water}
