% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanGenuchten.R
\name{vG}
\alias{vG}
\title{A Function}
\usage{
vG(P, alpha, n)
}
\arguments{
\item{parameters}{here.}
}
\description{
A Function
}
\examples{
hi
}
\keyword{conductivity}
\keyword{hydraulic}
\keyword{retention}
\keyword{saturated}
\keyword{water}
